requirements.txt
# AIX Data Graph Python Dependencies
# Core dependencies
requests>=2.28.0
paramiko>=3.0.0
pyyaml>=6.0
schedule>=1.2.0
python-dateutil>=2.8.2

# Database and time series
influxdb-client>=1.36.0
pandas>=1.5.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.16.0

# Web framework for API
flask>=2.3.0
flask-cors>=4.0.0

# Configuration and validation
pydantic>=2.0.0
click>=8.1.0

# Utilities
psutil>=5.9.0
cryptography>=41.0.0

